// frontend/it-support/apps/business/apis/models/ProjectStoreInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Business Service
 * business APIs
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface ProjectStoreInput
 */
export interface ProjectStoreInput {
  /**
   *
   * @type {string}
   * @memberof ProjectStoreInput
   */
  title: string;
  /**
   *
   * @type {string}
   * @memberof ProjectStoreInput
   */
  description: string;
  /**
   *
   * @type {Date}
   * @memberof ProjectStoreInput
   */
  startDate?: Date;
  /**
   *
   * @type {Date}
   * @memberof ProjectStoreInput
   */
  endDate?: Date;
  /**
   *
   * @type {number}
   * @memberof ProjectStoreInput
   */
  minBudget?: number;
  /**
   *
   * @type {number}
   * @memberof ProjectStoreInput
   */
  maxBudget?: number;
  /**
   *
   * @type {boolean}
   * @memberof ProjectStoreInput
   */
  isActive: boolean;
}

/**
 * Check if a given object implements the ProjectStoreInput interface.
 */
export function instanceOfProjectStoreInput(value: object): value is ProjectStoreInput {
  if (!("title" in value) || value["title"] === undefined) return false;
  if (!("description" in value) || value["description"] === undefined) return false;
  if (!("isActive" in value) || value["isActive"] === undefined) return false;
  return true;
}

export function ProjectStoreInputFromJSON(json: any): ProjectStoreInput {
  return ProjectStoreInputFromJSONTyped(json, false);
}

export function ProjectStoreInputFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ProjectStoreInput {
  if (json == null) {
    return json;
  }
  return {
    title: json["title"],
    description: json["description"],
    startDate: json["startDate"] == null ? undefined : new Date(json["startDate"]),
    endDate: json["endDate"] == null ? undefined : new Date(json["endDate"]),
    minBudget: json["minBudget"] == null ? undefined : json["minBudget"],
    maxBudget: json["maxBudget"] == null ? undefined : json["maxBudget"],
    isActive: json["isActive"],
  };
}

export function ProjectStoreInputToJSON(json: any): ProjectStoreInput {
  return ProjectStoreInputToJSONTyped(json, false);
}

export function ProjectStoreInputToJSONTyped(
  value?: ProjectStoreInput | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    title: value["title"],
    description: value["description"],
    startDate:
      value["startDate"] == null ? undefined : value["startDate"].toISOString().substring(0, 10),
    endDate: value["endDate"] == null ? undefined : value["endDate"].toISOString().substring(0, 10),
    minBudget: value["minBudget"],
    maxBudget: value["maxBudget"],
    isActive: value["isActive"],
  };
}